Modify the condition of setting git username and email
Previously, the git username and email were set as long as they were not
set on the target machine. Now we also check if these two variables are
defined or not. If not defined, then we don't set them; if defined, we
then check further whether they have already been set on the target
machine.

Therefore, if I don't want to change the git username and email on the
target machine, I only need to not define them when running the
playbook.