Install the latest version of VirtualBox by default
Previously, the `virtualbox.yml` always installed the VirtualBox from
the official Ubuntu package server. However, the version on the Ubuntu
package server may not be the latest. As a result, VirtualBox may pop up
notification to remind you of the new available version for update.

This commit alters the default behavior of the playbook to installing
the latest version of VirtualBox that's downloaded from the official
download page. When you get notified of the new available version, you
don't have to download the `.deb` package and install it manually. You
can just run this playbook to install it.

Closes: #17