Fix bugs
1). `COPY` by default makes the copied file be owned by `root`, even if
the current operator is already switched to a non-root user. Use
`--chown` to change the ownership.

2). I misused the variable `${HOME}` before which was probably not
defined as I thought. As a result, I had to copy the SSH key to a
temporary location first before copying it to the desired destination
location (`$HOME/.ssh/`). This is fixed by replacing `${HOME}` with
`/home/${USER_NAME}`.